events: Workaround badly behaving browser caching.
The code comments explain this issue in some detail.

In theory, the cache control headers that Zulip specifies via
patch_cache_control should prevent this situation from happening, but
we've had a number of reports of users seeing an invalid state clearly
caused by the browser using a stale version of the webapp's HTML and
thus page_params, resulting in it feeling like the Zulip window
travelled back in time a few hours (I've had it happen to me myself).

We can fix this by adding error checking that is reasonable to do in
any case (and could, in theory, catch other bugs): Rejecting requests
to fetch events from an events queue that have already been pruned /
garbage collected.

Fixes #12876.