subscriptions: Simplify the displaying of subs pane.
".stream-info-title" selectot is used to hide both
"#add_new_stream_title" and "#stream_settings_title"
classes. This will be helpful when we add new html
elements to display the risht side title of the
subscription overlay.

`clear_edit_panel` can be removed as the next line
where it is called makes it redundant, we only need
to de-select the stream row, as done in this commit.