message view: Don't mark selected message as read when narrowing.
For narrows where we select the first unread message,
on activation we have to select it. And on our initial
selection was also `force_rerender` the message list
which causes the message to get reselected.

When `MessageList.select_id` is called for the first time,
`previously_selected_id` is -1. However, because the rerender
function call is before the event trigger, `select_id` is
called again this time with `previously_selected_id` have the
same id as the to be selected message and this selection event
gets triggered before our first time's selection.

This above case only occurs only for narrow activation or
when a message is selected again by clicking on it (message
is already read for this case).

For the All Messages view we have similar logic where, since
opts does not contain `force_rerender` we pass `mark_read`
as false (which was previously unused) which skips the
"message_selected.zulip" event in `message_scroll.js`.

Fixes #10886.