message list data: Cache MLD objects.
We store the new MessageListData objects in the
`mld_cache` object, when a narrow is activated.

And if we are revisiting a narrow then we do not
fetch from the server again. The invalidation of
this cache will be implemented in subsequent
commits, thus we for now get incorrect data if
a message event changes our stored objects.